requirements.txt
flask
requests
gitpython
flask-cors
python-dotenv